# Scala: Improve and speed-up implementation