--- a/scala/src/main/scala/raytracer/BVH.scala
+++ b/scala/src/main/scala/raytracer/BVH.scala
@@ -1,8 +1,5 @@
 package raytracer
 import scala.collection.immutable.Nil
-import scala.concurrent.{Await, Future}
-import scala.concurrent.duration.Duration
-import scala.concurrent.ExecutionContext
 
 final case class AABB(min: Vec3, max: Vec3) {
   @inline def surroundingBox(that: AABB): AABB = {
@@ -43,30 +40,30 @@ final case class Leaf[A](aabb: AABB, a: A) extends BVH[A]
 final case class Split[A](aabb: AABB, left: BVH[A], right: BVH[A]) extends BVH[A]
 
 object BVH {
+  import java.util.concurrent.{Executors, ForkJoinPool, RecursiveTask}
   import scala.math.Ordering.Double.IeeeOrdering
 
-  def apply[A](f: A => AABB, allObjs: List[A])(implicit ec: ExecutionContext): BVH[A] = {
+  private final lazy val pool = Executors.newWorkStealingPool().asInstanceOf[ForkJoinPool]
+  sys.addShutdownHook(pool.shutdown())
 
-    def go(d: Int, n: Int, objs: List[A]): BVH[A] = objs match {
-      case Nil => throw new RuntimeException("BVH.apply: empty no nodes")
-      case x :: Nil => Leaf(f(x), x)
-      case xs => {
-        val (xsLeft, xsRight) =
-          xs.sortBy(a => f(a).axis(d))
-            .splitAt(n / 2)
-        def doLeft() = go(d+1, (n/2), xsLeft)
-        def doRight() = go(d+1, n-n/2, xsRight)
-        val (left, right) = if(n < 100) {
-          (doLeft(), doRight())
-        } else {
-          val l = Future { doLeft() }
-          val r = Future { doRight() }
-          (Await.result(l, Duration.Inf), Await.result(r, Duration.Inf))
-        }
-        val box = left.getAABB.surroundingBox(right.getAABB)
-        Split(box, left, right)
+  def apply[A](f: A => AABB, allObjs: List[A]): BVH[A] = {
+    class Go(d: Int, n: Int, objs: List[A]) extends RecursiveTask[BVH[A]] {
+      override def compute(): BVH[A] = objs match {
+        case Nil => throw new RuntimeException("BVH.apply: empty no nodes")
+        case x :: Nil => Leaf(f(x), x)
+        case xs =>
+          val (xsLeft, xsRight) = xs.sortBy(a => f(a).axis(d)).splitAt(n / 2)
+          def doLeft() = new Go(d+1, n/2, xsLeft)
+          def doRight() = new Go(d+1, n-n/2, xsRight)
+          val l = doLeft().fork()
+          val r = doRight().fork()
+          val left = l.join()
+          val right = r.join()
+          val box = left.getAABB.surroundingBox(right.getAABB)
+          Split(box, left, right)
       }
     }
-    go(0, allObjs.length, allObjs)
+
+    pool.invoke(new Go(0, allObjs.length, allObjs))
   }
 }