--- a/scala/src/main/scala/raytracer/Main.scala
+++ b/scala/src/main/scala/raytracer/Main.scala
@@ -1,7 +1,6 @@
 package raytracer
 
 import scala.util.Try
-import scala.concurrent.ExecutionContext.Implicits.global
 import raytracer.Raytracer.render
 
 object Main {