--- a/scala/src/main/scala/raytracer/Scene.scala
+++ b/scala/src/main/scala/raytracer/Scene.scala
@@ -1,10 +1,9 @@
 package raytracer
 
 import Raytracer.{Pos, Sphere, Camera, Objs}
-import scala.concurrent.ExecutionContext
 
 final case class Scene(camLookFrom: Pos, camLookAt: Pos, camFov: Double, spheres: List[Sphere]) {
-  def toObjsCam(width: Int, height: Int)(implicit ec: ExecutionContext): (Objs, Camera) =
+  def toObjsCam(width: Int, height: Int): (Objs, Camera) =
     (BVH(_.aabb, spheres), Camera(camLookFrom, camLookAt, Vec3(0,1,0), camFov, width.toDouble / height.toDouble))
 }