This PR contains three different changes which speed up the scala implementation (@athas If necessary, I can split it up in three different PRs):

* Rewrite saving image
* Optimize BVH by using its own thread-pool
* Minor optimizations, like avoiding duplicate calculations

Using `sbt bench` I can see a very nice boost in performance (Intel(R) Core(TM) i7-8650U CPU).
Before:
```
[info] Bench.construct   rgbbox  avgt    2     0.389          ms/op
[info] Bench.construct    irreg  avgt    2    15.754          ms/op
[info] Bench.render      rgbbox  avgt    2  1663.411          ms/op
[info] Bench.render       irreg  avgt    2   554.984          ms/op
```
After:
```
[info] Bench.construct   rgbbox  avgt    2     0.308          ms/op
[info] Bench.construct    irreg  avgt    2     8.974          ms/op
[info] Bench.render      rgbbox  avgt    2  1376.377          ms/op
[info] Bench.render       irreg  avgt    2   439.682          ms/op
```

A couple of words regarding the BVH rewrite. The default execution context is not designed for long lasting blocking operations, see [The Global Execution Context
](https://docs.scala-lang.org/overviews/core/futures.html#the-global-execution-context) for more information. In this case, every call of `Future` spawns a new thread.

As recommended, the new implementation uses a dedicated thread-pool (amount of threads = number of processors. The visualize the issue, here a couple of metrics.

Current implementation:
![threads-irreg-construct1](https://user-images.githubusercontent.com/5256214/80468838-03082880-8940-11ea-9e1c-6b0a01413a57.png)

![threads-irreg-construct2](https://user-images.githubusercontent.com/5256214/80469139-727e1800-8940-11ea-9ef7-386499b21717.png)

Spawns up multiple threads (in my case up to 20) on a 8-core machine, which leads to a low load in total (~35%).

New implementation:
![new-threads-irreg-construct1](https://user-images.githubusercontent.com/5256214/80469234-90e41380-8940-11ea-8999-55e1a9f1014c.png)
![new-threads-irreg-construct2](https://user-images.githubusercontent.com/5256214/80469253-96d9f480-8940-11ea-932e-3042386c379f.png)

The load is much higher, up to 80%.